feat(REC): mejoras en auditoria de usuarios (#89)
